+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/AlbumTagPanel.java
@@ -53,21 +53,21 @@ final class AlbumTagPanel extends JPanel {
 	private final State movingTags = State.state(false);
 	private final Control addTagControl = Control.builder()
 					.command(this::addTag)
-					.smallIcon(ICONS.get(Foundation.PLUS))
+					.icon(ICONS.get(Foundation.PLUS))
 					.build();
 	private final Control removeTagControl = Control.builder()
 					.command(this::removeTag)
-					.smallIcon(ICONS.get(Foundation.MINUS))
+					.icon(ICONS.get(Foundation.MINUS))
 					.enabled(selectionEmpty.not())
 					.build();
 	private final Control moveSelectionUpControl = Control.builder()
 					.command(this::moveSelectedTagsUp)
-					.smallIcon(ICONS.up())
+					.icon(ICONS.up())
 					.enabled(selectionEmpty.not())
 					.build();
 	private final Control moveSelectionDownControl = Control.builder()
 					.command(this::moveSelectedTagsDown)
-					.smallIcon(ICONS.down())
+					.icon(ICONS.down())
 					.enabled(selectionEmpty.not())
 					.build();