+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/CoverArtPanel.java
@@ -85,13 +85,13 @@ final class CoverArtPanel extends JPanel {
 		this.addButton = button()
 						.control(Control.builder()
 										.command(this::addCover)
-										.smallIcon(ICONS.get(Foundation.PLUS)))
+										.icon(ICONS.get(Foundation.PLUS)))
 						.transferFocusOnEnter(true)
 						.build();
 		this.removeButton = button()
 						.control(Control.builder()
 										.command(this::removeCover)
-										.smallIcon(ICONS.get(Foundation.MINUS)))
+										.icon(ICONS.get(Foundation.MINUS)))
 						.transferFocusOnEnter(true)
 						.enabled(imageSelected)
 						.build();