+++ demos/employees/src/main/java/is/codion/demos/employees/ui/DepartmentTablePanel.java
@@ -51,7 +51,7 @@ public class DepartmentTablePanel extends EntityTablePanel {
 		control(PRINT).set(Control.builder()
 						.command(this::viewEmployeeReport)
 						.caption("Employee Report")
-						.smallIcon(FrameworkIcons.instance().print())
+						.icon(FrameworkIcons.instance().print())
 						.enabled(tableModel().selection().empty().not())
 						.build());
 	}