+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/icon/package-info.java
@@ -19,7 +19,6 @@
 /**
  * Package configuration values:
  * <ul>
- * <li>{@link is.codion.swing.common.ui.icon.Icons#SIZE}
  * <li>{@link is.codion.swing.common.ui.icon.Icons#COLOR}
  * </ul>
  */